created sep log_reg file
